improve handling of logout
