+++ src/pages/FlowPage.tsx
@@ -167,7 +167,7 @@ export const FlowPage = () => {
   const [isEndingSession, setIsEndingSession] = useState(false)
   const [difficulty, setDifficulty] = useState<Difficulty>()
   const [player, setPlayer] = useState<Spotify.Player | null>(null)
-  const [deviceId, setDeviceId] = useState<string>('')
+  const [spotifyDeviceId, setSpotifyDeviceId] = useState<string>('')
   const [isPlaying, setIsPlaying] = useState(false)
   const [currentTrack, setCurrentTrack] = useState<{
     name: string
@@ -212,7 +212,7 @@ export const FlowPage = () => {
 
   useEffect(() => {
     const handleSessionComplete = async () => {
-      if (player && deviceId) {
+      if (player && spotifyDeviceId) {
         await player.pause()
         await SpotifyApiService.transferPlaybackToComputerDevice()
         player.disconnect() // Disconnect the Web Playback SDK player
@@ -228,7 +228,7 @@ export const FlowPage = () => {
     return () => {
       window.removeEventListener('flowSessionComplete', handleSessionComplete)
     }
-  }, [flowSession, player, deviceId])
+  }, [flowSession, player, spotifyDeviceId])
 
   useEffect(() => {
     const initSpotify = async () => {
@@ -242,7 +242,7 @@ export const FlowPage = () => {
         const newPlayer = await SpotifyApiService.createPlayer()
 
         newPlayer.addListener('ready', ({ device_id }: { device_id: string }) => {
-          setDeviceId(device_id)
+          setSpotifyDeviceId(device_id)
           const state = window.history.state?.usr
           const playlistToUse = state?.selectedPlaylist || localStorage.getItem('lastPlaylist')
           if (playlistToUse) {
@@ -350,7 +350,7 @@ export const FlowPage = () => {
     setIsEndingSession(true)
 
     // Stop playback, transfer to computer device, and clear player state
-    if (player && deviceId) {
+    if (player && spotifyDeviceId) {
       await player.pause()
       await SpotifyApiService.transferPlaybackToComputerDevice()
       player.disconnect() // Disconnect the Web Playback SDK player
@@ -416,13 +416,13 @@ export const FlowPage = () => {
   }
 
   const handlePlaylistChange = async (playlistId: string) => {
-    if (!deviceId) return
+    if (!spotifyDeviceId) return
     setSelectedPlaylistId(playlistId)
 
     // Always save the selected playlist to localStorage
     localStorage.setItem('lastPlaylist', playlistId)
     
-    await SpotifyApiService.startPlayback(playlistId, deviceId)
+    await SpotifyApiService.startPlayback(playlistId, spotifyDeviceId)
   }
 
   const MusicPlayer = () => (
